Fix popularity type errors across admin pages (#311)
* Fix TypeError in DataQualityPage when popularity is string

PostgreSQL can return numeric types as strings, causing `.toFixed()`
to fail with "n.toFixed is not a function" error. The page would
render blank due to unhandled runtime error.

Changed `actor.popularity?.toFixed(1)` to
`actor.popularity != null ? Number(actor.popularity).toFixed(1) : "N/A"`
to handle both null values and string-typed numbers.



* Cast popularity to float in all admin API queries

PostgreSQL's numeric(10,3) type is returned as a string by node-pg,
causing frontend .toFixed() calls to fail. Rather than adding
Number() wrappers on every frontend usage, fix at the source by
casting popularity::float in SQL queries.

This fixes the ActorManagementPage crash when filtering by popularity,
and preemptively fixes the same issue in:
- ActorDiagnosticPage
- HighPriorityActorsPage
- CacheManagementPage
- Various enrichment pages



* Address Copilot review feedback

- Simplify DataQualityPage popularity formatting (backend now returns numbers)
- Refactor cache.ts to use parameterized query instead of string interpolation
- Add regression tests for popularity::float cast in SQL queries



---------

+++ server/src/lib/db/admin-coverage-queries.ts
@@ -192,13 +192,14 @@ export async function getActorsForCoverage(
 
   // Use window function to get total count in same query (performance optimization)
   // Eliminates separate COUNT query - significant speedup for large tables
+  // Cast popularity to float to ensure JavaScript number type (pg returns numeric as string)
   const dataResult = await pool.query<ActorCoverageInfo & { total_count: string }>(
     `SELECT
        id,
        name,
        tmdb_id,
        deathday,
-       popularity,
+       popularity::float,
        has_detailed_death_info,
        enriched_at,
        age_at_death,
@@ -283,13 +284,14 @@ export async function getEnrichmentCandidates(
   minPopularity: number = 5,
   limit: number = 100
 ): Promise<ActorCoverageInfo[]> {
+  // Cast popularity to float to ensure JavaScript number type (pg returns numeric as string)
   const result = await pool.query<ActorCoverageInfo>(
     `SELECT
        id,
        name,
        tmdb_id,
        deathday,
-       popularity,
+       popularity::float,
        has_detailed_death_info,
        enriched_at,
        age_at_death,
